+++ ProsjektOppgaveWebAPI.EntityFramework/Repository/IGenericRepository.cs
@@ -6,6 +6,6 @@ public interface IGenericRepository<T> where T : class
     Task Update(T entity);
     Task Delete(T entity);
     
-    Task<T> GetById(long id);
+    Task<T> GetById(int id);
     IQueryable<T> GetAll();
 }
